show module name, format code
